move PeripheralManager handler to proxy
